+++ yggdrasil_decision_forests/learner/gradient_boosted_trees/gradient_boosted_trees.cc
@@ -2042,7 +2042,8 @@ GradientBoostedTreesLearner::GetGenericHyperParameterSpecification() const {
       "decision trees trained sequentially. Each tree is trained to predict "
       "and then \"correct\" for the errors of the previously trained trees "
       "(more precisely each tree predict the gradient of the loss relative to "
-      "the model output).");
+      "the model output).\nGBTs use [early stopping](early_stopping.md) to "
+      "avoid overfitting.");
 
   const auto proto_path =
       "learner/gradient_boosted_trees/gradient_boosted_trees.proto";