+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## HEAD
+
+### Features
+
+-   Early stopping is no longer triggered during first iterations. The initial
+    iteration for early stopping can be controlled with the new parameter
+    `early_stopping_initial_iteration` in `gradient_boosted_trees.proto`.
+
 ## 1.0.0 - 2022-09-07
 
 ### Features